Created 2d movement and rotation based on the WASD keys
The player object now moves in 4 directions following the WASD keys. The player also rotates to face the direction it is moving in. Additionally, I have added a circle to the front of the game object so that you can see which part of it is the front, as well as a wall to test that collision is working