Added memento objects
Created memento objects. When the player gets close to them and presses E, an event will be sent out that specifies the specific memento that was found. Right now the only thing that looks at it is the mementos themselves, but its easy to have a game manager listen to them as well.