Gave the ritual object full functionality
The ritual object now decides if you have enough mementos, and if you do, it will take you to the credits. If you do not, it will display a UI message telling you how many mementos you still need to find. In addition, I adjusted the UI canvas so that it always shows, and isn't hidden behind the environment